requirements.txt
Flask~=1.1.2
Flask-RESTful
SQLAlchemy< 1.4.0
Flask-Migrate
alembic~=1.5.8
sqlalchemy-utils
psycopg2
requests~=2.25.1
Flask-Script==2.0.6
odfpy==1.3.4
jsonschema==3.2.0
python-dotenv~=0.17.0

